feat(tiktok): 新增数据开放服务音乐榜相关接口

+++ src/SKIT.FlurlHttpClient.ByteDance.TikTok/Models/DataExternalBillboard/Sport/DataExternalBillboardSportOverallResponse.cs
@@ -94,7 +94,7 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("hot_value")]
                         [System.Text.Json.Serialization.JsonPropertyName("hot_value")]
-                        public int HotValue { get; set; }
+                        public double HotValue { get; set; }
 
                         /// <summary>
                         /// 获取或设置视频列表。
